* Tagged 0.3 beta 1 release: - Compiler warning fix - Player and Global Storage is now string, but fully compatible with integers - Updated databasemanager, schemas and recreated sqlite database
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@671 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ connection.cpp
@@ -83,7 +83,7 @@ bool ConnectionManager::isDisabled(uint32_t clientIp)
 	IpConnectionMap::const_iterator it = ipConnectionMap.find(clientIp);
 	if(it != ipConnectionMap.end())
 	{
-		if((it->second.loginsAmount >= maxLoginTries) && (time(NULL) < it->second.lastLogin + loginTimeout))
+		if((it->second.loginsAmount >= maxLoginTries) && ((uint64_t)time(NULL) < it->second.lastLogin + loginTimeout))
 			return true;
 	}
